### 2025-01-16 #7676 @editorialbot
[PRE REVIEW]: SOFICS: Optimization Framework for Structures under Detonation-Induced Fluid Flows
**Submitting author:** @anarkh97 (<a href="http://orcid.org/0009-0004-4281-611X">Aditya Narkhede</a>)
**Repository:** https://github.com/anarkh97/sofics
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
