### 2025-01-27 #7676 @anarkh97
Hi,

Thank you for taking the time to consider our paper. I have a quick question: Can we recommend reviewers who are not included in the JOSS reviewer list, or are we required to choose from the list?

Thanks in advance for your clarification!
