+++ paper/paper.bib
@@ -31,7 +31,8 @@
 	title={Dakota, A Multilevel Parallel Object-Oriented Framework for Design Optimization, Parameter Estimation, Uncertainty Quantification, and Sensitivity Analysis: Version 6.16 User’s Manual},
 	author={Adams, Brian M and Bohnhoff, William J and Dalbey, Keith R and Ebeida, Mohamed S and Eddy, John P and Eldred, Michael S and Hooper, Russell W and Hough, Patricia D and Hu, Kenneth T and Jakeman, John D and others},
 	year={2022},
-	institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)}
+	institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)},
+	url={https://doi.org/10.2172/1868142},
 }
 @article{farhat_robust_2010,
 	title = {Robust and provably second-order explicit-explicit and implicit-explicit staggered time-integrators for highly non-linear compressible fluid-structure interaction problems},
@@ -75,6 +76,7 @@
 	author = {Ma, Wentao and Zhao, Xuning and Gilbert, Christine and Wang, Kevin},
 	year = {2022},
 	pages = {111527},
+	url={https://doi.org/10.1016/j.ijsolstr.2022.111527},
 }
 @article{farhat_dynamic_2013,
 	title={Dynamic implosion of underwater cylindrical shells: experiments and computations},
@@ -84,7 +86,8 @@
 	number={19},
 	pages={2943--2961},
 	year={2013},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	url={https://doi.org/10.1016/j.ijsolstr.2013.05.006},
 }
 @article{zhao_simulating_2023,
 	title={Simulating laser-fluid coupling and laser-induced cavitation using embedded boundary and level set methods},
@@ -93,7 +96,8 @@
 	volume={472},
 	pages={111656},
 	year={2023},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	url={https://doi.org/10.1016/j.jcp.2022.111656},
 }
 
 @article{islam_plasma_2023,
@@ -103,7 +107,8 @@
 	volume={58},
 	pages={101927},
 	year={2023},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	url={https://doi.org/10.1016/j.eml.2022.101927},
 }
 @article{islam_fluid_2023,
 	title={Fluid-solid coupled simulation of hypervelocity impact and plasma formation},
@@ -111,13 +116,17 @@
 	journal={International Journal of Impact Engineering},
 	pages={104695},
 	year={2023},
-	publisher={Elsevier}
+	publisher={Elsevier},
+	url={https://doi.org/10.1016/j.ijimpeng.2023.104695},
 }
-@article{zhao_long_2023,
-	title={Long-Pulse Laser-Induced Cavitation: A Race Between Advection and Phase Transition},
-	author={Zhao, Xuning and Ma, Wentao and Chen, Junqin and Xiang, Gaoming and Zhong, Pei and Wang, Kevin},
-	journal={arXiv preprint arXiv:2308.11866},
-	year={2023}
+@article{zhao_long_2023, 
+	title={Vapour bubbles produced by long-pulsed laser: a race between advection and phase transition}, 
+	volume={999}, 
+	doi={10.1017/jfm.2024.989}, 
+	journal={Journal of Fluid Mechanics}, 
+	author={Zhao, Xuning and Ma, Wentao and Chen, Junqin and Xiang, Gaoming and Zhong, Pei and Wang, Kevin}, 
+	year={2024}, 
+	pages={A103}
 }
 @article{cao_shock_2021,
[…]
