### 2025-02-06 325fcf6
included doi field in bib file.
