+++ paper/paper.md
@@ -41,7 +41,7 @@ We develop `SOFICS`[^1], shown in \autoref{fig:sofics}, which streamlines the op
 [^2]: https://github.com/kevinwgy/m2c
 [^3]: https://github.com/anarkh97/aero-s-foam
 
-The advantages of using `M2C` for design optimization are three-fold. First, `M2C` is directly coupled with our open-source, finite element-based structural dynamics solver `Aero-S`[^3] using the partitioned procedure [@farhat_robust_2010]. Both solvers are designed for parallel computations, utilizing the Message Passing Interface (MPI) [-@mpi41] for efficient distributed processing. Second, we use the embedded boundary method [@ma_computational_2022;@wang_computational_2012;@wang_computational_2015] to accommodate large deflections. The method tracks the fluid-structure interface explicitly within a fixed, non-uniform, non-body conforming finite volume mesh, eliminating the need for mesh motion and frequent re-meshing. As a result, a single, fixed fluid mesh can be computed *a priori* and used consistently across different design configurations. Finally, `M2C` computes the interfacial mass, momentum, and energy fluxes across the fluid-structure interface using the **FI**nite **V**olume method with **E**xact multi-material **R**iemann problems (FIVER), which solves a 1D fluid-structure Riemann problem locally featuring a constant interface velocity obtained from the structural solver [@ma_computational_2022;@wang_computational_2012]. The coupled fluid-structure simulation model has been validated for several multiphase flow and fluid-structure interaction problems. These include underwater and aerial explosions and implosions [@farhat_dynamic_2013;@ma_computational_2022,@narkhede_fluid_2025], cavitation [@cao_shock_2021;@zhao_long_2023], hypervelocity impacts [@islam_fluid_2023;@islam_plasma_2023], and shockwave and laser lithotripsy [@cao_shock_2019;@wang_multiphase_2017;@zhao_simulating_2023].
+The advantages of using `M2C` for design optimization are three-fold. First, `M2C` is directly coupled with our open-source, finite element-based structural dynamics solver `Aero-S`[^3] using the partitioned procedure [@farhat_robust_2010]. Both solvers are designed for parallel computations, utilizing the Message Passing Interface (MPI) [-@mpi41] for efficient distributed processing. Second, we use the embedded boundary method [@ma_computational_2022;@wang_computational_2012;@wang_computational_2015] to accommodate large deflections. The method tracks the fluid-structure interface explicitly within a fixed, non-uniform, non-body conforming finite volume mesh, eliminating the need for mesh motion and frequent re-meshing. As a result, a single, fixed fluid mesh can be computed *a priori* and used consistently across different design configurations. Finally, `M2C` computes the interfacial mass, momentum, and energy fluxes across the fluid-structure interface using the **FI**nite **V**olume method with **E**xact multi-material **R**iemann problems (FIVER), which solves a 1D fluid-structure Riemann problem locally featuring a constant interface velocity obtained from the structural solver [@ma_computational_2022;@wang_computational_2012]. The coupled fluid-structure simulation model has been validated for several multiphase flow and fluid-structure interaction problems. These include underwater and aerial explosions and implosions [@farhat_dynamic_2013;@ma_computational_2022;@narkhede_fluid_2025], cavitation [@cao_shock_2021;@zhao_long_2023], hypervelocity impacts [@islam_fluid_2023;@islam_plasma_2023], and shockwave and laser lithotripsy [@cao_shock_2019;@wang_multiphase_2017;@zhao_simulating_2023].
 
[…]
