+++ paper/paper.bib
@@ -6,7 +6,9 @@
 	number={6},
 	pages={620--631},
 	year={1998},
-	publisher={American Institute of Physics}
+	publisher={American Institute of Physics},
+	doi={10.1063/1.168744},
+	url={https://doi.org/10.1063/1.168744},
 }
 @article{alnaes_fenics_2015,
 	title     = {The {FEniCS} Project Version 1.5},
@@ -15,6 +17,7 @@
 	year      = {2015},
 	volume    = {3},
 	doi       = {10.11588/ans.2015.100.20553},
+	url       = {https://doi.org/10.11588/ans.2015.100.20553}
 }
 @article{chourdakis_precice_2022,
 	author = {Chourdakis, G and Davis, K and Rodenberg, B and Schulte, M and Simonis, F and Uekermann, B and Abrams, G and Bungartz, HJ and Cheung Yau, L and Desai, I and Eder, K and Hertrich, R and Lindner, F and Rusch, A and Sashko, D and Schneider, D and Totounferoush, A and Volland, D and Vollmer, P and Koseomur, OZ},
@@ -33,12 +36,14 @@
 	year={2022},
 	institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)},
 	url={https://doi.org/10.2172/1868142},
+	doi={10.2172/1868142},
 }
 @article{farhat_robust_2010,
 	title = {Robust and provably second-order explicit-explicit and implicit-explicit staggered time-integrators for highly non-linear compressible fluid-structure interaction problems},
 	volume = {84},
 	issn = {0029-5981},
 	doi = {10.1002/nme.2883},
+	url={https://doi.org/10.1002/nme.2883},
 	number = {1},
 	journal = {International Journal for Numerical Methods in Engineering},
 	author = {Farhat, C. and Rallu, A. and Wang, K. and Belytschko, T.},
@@ -51,6 +56,7 @@
 	volume = {70},
 	issn = {0271-2091},
 	doi = {10.1002/fld.3659},
+	url={https://doi.org/10.1002/fld.3659},
 	number = {4},
 	journal = {International Journal for Numerical Methods in Fluids},
 	author = {Wang, K. and Grétarsson, J. and Main, A. and Farhat, C.},
@@ -67,6 +73,8 @@
 	author = {Wang, K. G. and Lea, P. and Farhat, C.},
 	year = {2015},
 	pages = {585--623},
+	url={https://doi.org/10.1002/nme.4873},
+	doi={10.1002/nme.4873},
 }
 @article{ma_computational_2022,
 	title = {Computational analysis of bubble–structure interactions in near-field underwater explosion},
@@ -77,6 +85,7 @@
 	year = {2022},
 	pages = {111527},
 	url={https://doi.org/10.1016/j.ijsolstr.2022.111527},
+	doi={10.1016/j.ijsolstr.2022.111527},
 }
 @article{farhat_dynamic_2013,
 	title={Dynamic implosion of underwater cylindrical shells: experiments and computations},
@@ -88,6 +97,7 @@
 	year={2013},
 	publisher={Elsevier},
 	url={https://doi.org/10.1016/j.ijsolstr.2013.05.006},
+	doi={10.1016/j.ijsolstr.2013.05.006},
 }
 @article{zhao_simulating_2023,
 	title={Simulating laser-fluid coupling and laser-induced cavitation using embedded boundary and level set methods},
@@ -98,6 +108,7 @@
 	year={2023},
 	publisher={Elsevier},
[…]
