### 2025-03-20 a12745a
1. SOFICS modified for ADOPT. 2. Driver takes in command line options. 3. Pre- and Post-processing script can be provided    to the driver through command line options.
