+++ src/checks.sh
@@ -6,76 +6,129 @@
 BUILD_DIR=$(dirname $DRIVER_DIR)
 PACKAGE_DIR=$BUILD_DIR/packages
 
-if [[ $TEMPLATE_DIR == "" ]]; then
-  printf "*** Warning: User did not provide a template directory. Using "
-  printf "the default, which is the directory in which Dakota was executed.\n"
+if [ -z "$TEMPLATE_DIR" ]; then
+  printf "*** Warning: User did not provide a template "
+  printf "directory. Using the default, which is the "
+  printf "directory in which Dakota was executed.\n"
   TEMPLATE_DIR=$LAUNCH_DIR
 fi
 
-if [[ $GMSH_EXE == "" ]]; then
-  printf "*** Error: Executable for GMSH (FE mesher) was not provided. "
-  printf "Aborting ...\n"
-  #printf "*** Warning: Executable for GMSH (FE mesher) was not provided. "
-  #printf "Using default one ...\n"
-  #if [[ ! -e $PACKAGE_DIR/install/bin/gmsh ]]; then
-  #  printf "*** Error: No executable for solid solver was found. "
-  #  printf "Aborting ...\n"
-  #  exit 1
-  #fi
-  #GMSH_EXE=$PACKAGE_DIR/install/bin/gmsh
-fi
+#------------------------------------------------------------------------------
+# Executables checks. (Assign defaults whenever possible)
+#------------------------------------------------------------------------------
 
-if [[ $GMSH_INPUT == "" ]]; then
-  printf "*** Error: GMSH (FE mesher) input file was not provided. "
-  printf "Aborting ...\n"
-  exit 1
-fi
+# AN: Handled later.
+#if [ -z "$GMSH_EXE" ]; then
+#  printf "*** Error: Executable for GMSH (FE mesher) was not provided. "
+#  printf "Aborting ...\n"
+#  exit 1
+#  #printf "*** Warning: Executable for GMSH (FE mesher) was not provided. "
+#  #printf "Using default one ...\n"
+#  #if [[ ! -e $PACKAGE_DIR/install/bin/gmsh ]]; then
+#  #  printf "*** Error: No executable for solid solver was found. "
+#  #  printf "Aborting ...\n"
+#  #  exit 1
+#  #fi
+#  #GMSH_EXE=$PACKAGE_DIR/install/bin/gmsh
+#elif [ -n "$GMSH_EXE" ] && [[ ! -x "$GMSH_EXE" ]]; then
+#  printf "*** Error: \"${GMSH_EXE}\" is not a valid executable. "
+#  printf "Aborting ...\n"
+#  exit 1
+#fi
 
-if [[ $M2C_SIZE == "" ]]; then
-  printf "*** Error: Resource requirements for fluid solver were not "
-  printf "specified. Aborting ...\n"
-  exit 1
-fi
-
-if [[ $M2C_EXE == "" ]]; then
-  printf "*** Warning: Executable for the fluid solver was not provided. "
+if [ -z "$M2C_EXE" ]; then
+  printf "*** Warning: Executable for M2C was not provided. "
   printf "Using default one ...\n"
-  if [[ ! -e $PACKAGE_DIR/m2c/m2c ]]; then
-    printf "*** Error: No executable for fluid solver was found. "
+  if [[ ! -x $PACKAGE_DIR/m2c/m2c ]]; then
+    printf "*** Error: Could not find a valid executable for M2C. "
     printf "Aborting ...\n"
     exit 1
   fi
   M2C_EXE=$PACKAGE_DIR/m2c/m2c
-fi
-
-if [[ $M2C_INPUT == "" ]]; then
[…]
