+++ src/analysis_setup.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 #------------------------------------------------------------------------------
-# Perform checks before proceeding
+# Perform checks before proceeding (in CURRENT shell)
 #------------------------------------------------------------------------------
 if ! source "${DRIVER_DIR}/checks.sh"; then
   # NOTE: Here we do not write a results file and force dakota to fail.
@@ -12,29 +12,8 @@ if ! source "${DRIVER_DIR}/checks.sh"; then
 fi
 
 #------------------------------------------------------------------------------
-# Copy all templates to the working directory
-#------------------------------------------------------------------------------
-cp "$TEMPLATE_DIR/${GMSH_INPUT}.template" "$WORKING_DIR/$GMSH_INPUT"
-cp "$TEMPLATE_DIR/${AEROS_INPUT}.template" "$WORKING_DIR/$AEROS_INPUT"
-cp "$TEMPLATE_DIR/${M2C_INPUT}.template" "$WORKING_DIR/$M2C_INPUT"
-# copy any auxilary inputs that were provided for m2c
-if [ -n "$M2C_AUX" ]; then
-  # get names of all auxilary inputs
-  IFS=: read -r -a fluid_aux_inps <<< "$M2C_AUX"
-
-  # remove empty fields (or spaces)
-  fluid_aux_inps=("${fluid_aux_inps[@]// /}")
-
-  for i in "${!fluid_aux_inps[@]}"; do
-    cp "$TEMPLATE_DIR/${fluid_aux_inps[$i]}.template" \
-      "$WORKING_DIR/${fluid_aux_inps[$i]}"
-  done
-fi
-
-#------------------------------------------------------------------------------
 # Read Dakota Parameter file and forward variables to a pre-processor
 #------------------------------------------------------------------------------
-
 num_var=$(awk 'NR==1 {print $1}' "$WORKING_DIR/$DAK_PARAMS")
 # Dakota specifies continous design variables as derivative variables
 # even if gradient-free optimization is employed. There is no other 
@@ -54,15 +33,15 @@ fi
 # NOTE: dakota always writes the continuous design variables first.
 # TODO: Extend to include other kinds of design variables as well.
 cdv_list=""
-for i in $(seq 1 $num_cdv)
+for i in $(seq 1 "$num_cdv")
 do
-  val=$(awk -v line=$i 'NR==line+1 {print $1}' "$WORKING_DIR/$DAK_PARAMS")
-  desc=$(awk -v line=$i 'NR==line+1 {print $2}' "$WORKING_DIR/$DAK_PARAMS")
+  val=$(awk -v line="$i" 'NR==line+1 {print $1}' "$WORKING_DIR/$DAK_PARAMS")
+  desc=$(awk -v line="$i" 'NR==line+1 {print $2}' "$WORKING_DIR/$DAK_PARAMS")
 
   # This is to ensure that only continuous design variables type were
   # used in the Dakota input file. Will be extended in future and this condition
   # will be relaxed to accommodate discrete variables of type INTEGER and STRING. 
-  if [[ $desc != "cdv_$i" ]]; then
+  if [[ "$desc" != "cdv_$i" ]]; then
     printf "*** Error: This utility expects continuous design "
     printf "variables with dakota's default descriptors. Found a variable "
     printf "with descriptor %s. Aborting ...\n" "$desc"
@@ -139,16 +118,78 @@ do
 done
 
 #------------------------------------------------------------------------------
+# Perform FEST related checks and generate "META" file
[…]
