+++ src/pre_processor.sh
@@ -25,19 +25,14 @@ then
   csv=("${csv[@]// /}")
 fi
 
+#------------------------------------------------------------------------------
+# Sub-directory for structural mesh.
+#------------------------------------------------------------------------------
+
 # NOTE: Not strictly necessary, but it's conventional to place 
 # structure files in a separate directory for better organization.
 struct_dir=$WORKING_DIR/StructModel
 struct_inp=$WORKING_DIR/$AEROS_INPUT
-
-gmsh_name=$(basename "$WORKING_DIR/$GMSH_INPUT" .geo)
-gmsh_out="${gmsh_name}.msh"
-
-# local log files
-gmsh2aeros_log=$struct_dir/log.out
-
-# add evaluation number
-sed -i "s/{eval_num}/${DAK_EVAL_NUM}/g" $struct_inp
 
 # make the struct model directory
 if [ ! -d "${struct_dir}" ]
@@ -45,38 +40,62 @@ then
   mkdir "${struct_dir}"
 fi
 
+#------------------------------------------------------------------------------
+# Check GMSH input file and Executable. Move to sub-directory.
+#------------------------------------------------------------------------------
+
+if [[ ! -f "$TEMPLATE_DIR/${GMSH_INPUT}.template" ]]; then
+  printf "*** Error: Could not find a template file for Gmsh input "
+  printf "file (or template extension is missing). Required by default "
+  printf "structural pre-processor.\n"
+  exit 1
+fi
+if [ -z "$GMSH_EXE" ]; then
+  printf "*** Error: Executable for GMSH (FE mesher) was not provided. "
+  printf "Required by the default structural pre-processor. Aborting ...\n"
+  exit 1
+  #printf "*** Warning: Executable for GMSH (FE mesher) was not provided. "
+  #printf "Using default one ...\n"
+  #if [[ ! -e $PACKAGE_DIR/install/bin/gmsh ]]; then
+  #  printf "*** Error: No executable for solid solver was found. "
+  #  printf "Aborting ...\n"
+  #  exit 1
+  #fi
+  #GMSH_EXE=$PACKAGE_DIR/install/bin/gmsh
+elif [ -n "$GMSH_EXE" ] && [[ ! -x "$GMSH_EXE" ]]; then
+  printf "*** Error: \"%s\" is not a valid executable. " "${GMSH_EXE}"
+  printf "Required by the default structural pre-processor. Aborting ...\n"
+  exit 1
+fi
+
+cp "$TEMPLATE_DIR/${GMSH_INPUT}.template" "$struct_dir/$GMSH_INPUT"
+
+gmsh_name=$(basename "$struct_dir/$GMSH_INPUT" .geo)
+gmsh_out="${gmsh_name}.msh"
+# local log files
+gmsh2aeros_log="$struct_dir/log.out"
 
 #------------------------------------------------------------------------------
 # Update gmsh input file
 #------------------------------------------------------------------------------
 substitution_error=0
-if [[ ${#cdv[@]} == $num_cdv ]]; then
-  for i in "${!cdv[@]}"
-  do
-    if ! sed -i "s/{cdv_$((i+1))}/${cdv[$i]}/g" "$WORKING_DIR/$GMSH_INPUT"; then
-      printf "*** Error: Substitution for continuous design variable cdv_${i} "
-      printf "in file ${WORKING_DIR}/${GMSH_INPUT} failed.\n"
-      substitution_error=$((substitution_error+1))
-    fi
-  done
-else
-  printf "*** Error: Required number of design variables not found.\n"
-  exit 1
-fi
[…]
