### 2025-03-22 e5b90bd
Meta generator implemented. Testing. (work in progress)
