### 2025-03-22 f02d781
Meta file related variables included. (Work in progress)
