### 2025-04-10 fe8c2f5
1. Fixes for compatibility with ADOPT 2. Variables are placed under double quotations now.
