+++ src/error_driver.sh
@@ -50,7 +50,7 @@ printf "with process id %s.\033[0m\n" "${host_list[*]}" "$mpi_pid"
 # Wait for completion
 # -----------------------------------------------------------------------------
 
-wait $mpi_pid
+wait "$mpi_pid"
 
 # -----------------------------------------------------------------------------
 # Exit assuming everything was successfull
