### 2025-04-10 d68d0b5
More modifications for ADOPT.
Testing.
