### 2025-04-11 e42e464
Minor bug bixes for error evaluations.
