+++ src/driver.sh
@@ -100,7 +100,7 @@ then
     # add garbage value for MSE when performing TRUE/APPROX simulation
     # In case of TRUE simulation, these will be corrected in the error
     # simulation.
-    printf "0.0  MSE\n" >> "$WORKING_DIR/$DAK_RESULTS"
+    printf "    0.0  MSE\n" >> "$WORKING_DIR/$DAK_RESULTS"
 
   else
 
@@ -120,7 +120,7 @@ then
     fi
 
     for i in $(seq 1 "$num_funcs"); do
-      printf "0.0\n" >> "$WORKING_DIR/$DAK_RESULTS"
+      printf "    0.0\n" >> "$WORKING_DIR/$DAK_RESULTS"
     done
 
     NRMSE=$(
@@ -128,7 +128,7 @@ then
       cut -d ":" -f 2
     )
 
-    printf "%s  MSE\n" "$NRMSE" > "$WORKING_DIR/$DAK_RESULTS"
+    printf "    %s  MSE\n" "$NRMSE" > "$WORKING_DIR/$DAK_RESULTS"
   fi
 else
   # unsuccessfull evaluation
