### 2025-04-12 a548de0
Some more modifications for ADOPT.
