### 2025-04-14 b525008
M2C auxillary files do not need template extension.
