### 2025-04-18 52fe901
SOFICS compatible with latest changes in ADOPT.
