### 2025-04-23 #7676 @kyleniemeyer
@anarkh97 ok - in the meantime, I will ask a few existing JOSS reviewers.
