### 2025-04-23 #7676 @kyleniemeyer
Hi @corentin-dev and @Balaje would either or both of you be able to review this submission for JOSS? It relates to fluid-structure interaction, though with a focus on optimizing the structures. Thanks!
