### 2025-04-29 #7676 @corentin-dev
@kyleniemeyer yes, I can work on a review of the package.
