### 2025-04-29 #7676 @kyleniemeyer
Thanks @corentin-dev!
