### 2025-04-29 #7676 @kyleniemeyer
@Zhengyu-Huang would you be available to review this submission for JOSS, "SOFICS: Optimization Framework for Structures under Detonation-Induced Fluid Flows"?  If you haven't reviewed for us before, we use a [checklist-driven review process](https://joss.readthedocs.io/en/latest/review_checklist.html) that takes place in the open on GitHub, where the focus is primarily on the software itself.
