### 2025-06-23 26275a5
Timing code and build changes.
1. Driver now times the total elapsed time from sub-driver initiation to exit.
   - Addresses an issue on Tinkercliffs where evaluations could become zombies,
     requiring manual termination to continue optimization.
2. Updated bash build script to clearly display all shell files copied during
   the build process.
