+++ src/CMakeLists.txt
@@ -31,22 +31,30 @@ add_executable(postprocessor
 postprocessor.cpp
 )
 
-# copy bash scripts
-add_custom_target(sofics_scripts ALL)
-
 # Collect all .sh files in the source directory into a list
 file(GLOB BASH_FILES "${CMAKE_CURRENT_SOURCE_DIR}/*.sh")
 
 # Copy the collected files to the desired destination
 if(BASH_FILES)
+
+  set(SOFICS_SCRIPT_OUTPUTS "")
+
   foreach(shfile ${BASH_FILES})
     get_filename_component(fname ${shfile} NAME)
     add_custom_command(
-        TARGET sofics_scripts
-        COMMAND ${CMAKE_COMMAND} -E copy ${shfile} ${CMAKE_CURRENT_BINARY_DIR}/${fname}
-        DEPENDS ${shfile}
+      OUTPUT "${CMAKE_CURRENT_BINARY_DIR}/${fname}"
+      COMMAND ${CMAKE_COMMAND} -E copy_if_different ${shfile} "${CMAKE_CURRENT_BINARY_DIR}/${fname}"
+      DEPENDS ${shfile}
     )
+
+    list(APPEND SOFICS_SCRIPT_OUTPUTS "${CMAKE_CURRENT_BINARY_DIR}/${fname}")
   endforeach()
+
+  add_custom_target(sofics_scripts ALL
+    DEPENDS ${SOFICS_SCRIPT_OUTPUTS}
+    COMMENT ""
+  )
+
 else()
   message(ERROR "The SOFICS bash files not found.")
 endif()
