### 2025-06-24 #7676 @kyleniemeyer
Hello @chennachaos and @albansouche, would either (or both) of you be able to review this submission ("SOFICS: Optimization Framework for Structures under Detonation-Induced Fluid Flows") for JOSS? Thanks!
