### 2025-06-26 #8494 @kyleniemeyer
👋 @anarkh97 @corentin-dev @albansouche the actual review will take place in this issue.

@corentin-dev and @albansouche, you can generate your individual reviewer checklists with `@editorialbot generate my checklist`; additional comments (either here or issues in the software repo) are of course welcome. Thank you!
