### 2025-07-10 c4aa380
Expanding local scheduling to STATIC and DYNAMIC
