+++ src/error_driver.sh
@@ -4,35 +4,13 @@
 # Calculate node (list) for the current evaluation.
 #------------------------------------------------------------------------------
 
-# find correct relative node position
-task_per_node=$(
-  echo "$SLURM_TASKS_PER_NODE" |
-  sed 's|^[^0-9]*\([0-9]\+\).*|\1|'
-)
-total_proc=$((
-  "$FEST_SIZE"
-))
-applic_nodes=$((
-  ("$total_proc"+"$task_per_node"-1) / 
-  "$task_per_node"
-))
-relative_eval_num=$((
-  ("$DAK_EVAL_NUM"-1) % "$EVALUATION_CONCURRENCY"
-))
-relative_node=$((
-  ( "$relative_eval_num" * "$total_proc" ) / "$task_per_node"
-))
-
-# get all nodes assigned to dakota process
-node_list=(`scontrol show hostnames`)
-
-# get host list for this process
-host_list="${node_list[$relative_node]}"
-for i in $(seq 1 $((applic_nodes - 1)))
-do
-  host_list="$host_list,
-  ${node_list[$((relative_node+i))]}"
-done
+host_list=""
+total_proc=1
+if ! source "${DRIVER_DIR}/node_list_calculator.sh" "$host_list" "0" "1"
+then
+  printf "*** Error: Could not calculate the list of host nodes "
+  printf "for (error) evaluation %s.\n" "${DAK_EVAL_NUM}"
+fi
 
 ### run analysis in background
 mpiexec --bind-to none \
@@ -43,7 +21,7 @@ mpiexec --bind-to none \
 
 mpi_pid=$!
 
-printf "\033[34mLaunching Evaluation %s on nodes " "${DAK_EVAL_NUM}"
+printf "\033[34mLaunching (error) Evaluation %s on nodes " "${DAK_EVAL_NUM}"
 printf "%s with process id %s.\033[0m\n" "${host_list[*]}" "$mpi_pid"
 
 # -----------------------------------------------------------------------------
