### 2025-07-30 #8494 @kyleniemeyer
These are good questions, @albansouche - and you are correct that ~400-500 source lines of code would not be sufficient for JOSS. However, @anarkh97, if `m2c` is meant to be part of this package—rather than an external library—then that would change the situation.

I fully agree that a more-formal build/installation script would be helpful here, rather than a list of instructions in the README for a relatively complex process with multiple dependencies.
