### 2025-08-10 #8494 @anarkh97
Dear @albansouche,

Thank you for your patience. We sincerely appreciate your insightful comments and constructive suggestions. In response to the review, we have carefully revised a few things to address each point raised. Below, we provide our detailed responses to each comment and suggestion, ensuring transparency in how the feedback has been (or will be) incorporated. Please feel free to let us know if you have any further questions or additional suggestions.

Regards,
Aditya

## Comment #1

***I counted 487 lines of bash scripts in the sofics source code, which seems limited for JOSS publication. From the manuscript and repository, it seems m2c is the actual software while sofics mainly acts as a wrapper connecting different libraries. Is this correct? If so, why is m2c itself not the software being proposed for publication (<https://github.com/kevinwgy/m2c>)?***

**Response:** We sincerely appreciate the reviewer's thoughtful comment and questions. While the bash scripts play a significant role, they are only part of the efforts and contributions reflected in this publication. The development of SOFICS also involved extending the M2C solver to support two-way fluid-structure coupling, model the thermodynamics of gaseous detonation products, and support the automated input/output flow for structural optimization. On the other hand, we chose to propose SOFICS (rather than M2C) in this publication because it represents a multi-component software framework tailored for a specific, impactful application area: structural optimization under explosive loading. In contrast, M2C is a more general-purpose computational fluid dynamics (CFD) code with broader applicability that goes beyond SOFICS.

Below, we try to provide some details in support of the argument above.

A. The M2C extensions (source code: [***https://github.com/kevinwgy/m2c***](https://github.com/kevinwgy/m2c)) developed specifically for SOFICS include:

1.  Support of two-way coupling with concurrently running programs, including a specialization for the Aero-S solver to support fluid-structure interaction (FSI) simulations. This capability is mainly implemented in ConcurrentProgramsHandler.h/cpp and AerosMessenger.h/cpp, with support logic in the main solver loop.

2.  Jones-Wilkins-Lee (JWL) equation of state (EOS): This EOS was implemented to model the complex and highly nonlinear thermodynamic behavior of detonation products, which is critical for accurately capturing the detonation-induced loads. Unlike ideal gas or stiffened gas models, the JWL EOS is specifically formulated to describe the high-pressure, high-energy expansion of the gaseous explosion products expanding outwards. The core functions of the JWL EOS were implemented in VarFcnJWL.h. Additional changes were made across the codebase to support this capability, including Riemann solvers ExactRiemannSolverBase.h/cpp), flux functions (FluxFcnGenRoe.h, FluxFcnHLLC.h), and I/O routines (IoData.h/cpp).

3.  UserDataFile input option: This option was implemented to support the initialization of detonation-induced loading scenarios. It
    allows users to prescribe detailed initial conditions such as the explosive region geometry, and the initial variation of thermodynamic quantities within the explosive region, which are essential for simulating blast wave generation and propagation. This feature was mostly implemented in UserDefinedState.h/cpp, SpaceInitializer.h/cpp, and LevelSetOperator.h/cpp.

B. The SOFICS framework combines the following standalone software, each playing a different role.

1.  M2C: A high-resolution compressible CFD solver, used in SOFICS for simulating the explosive-induced fluid flow and predicting the transient pressure load on the structure.

2.  Aero-S: A structural dynamics solver based on finite element methods, used in SOFICS for simulating the deformation and failure of the solid structure being optimized.

3.  Dakota: An open-source optimization toolkit developed by Sandia National Laboratories, used in SOFICS for directing the iterative optimization process (which involves repeated calls to M2C and Aero-S for simulations with different inputs).

4.  Gmsh: An open-source mesh generation software, used in SOFICS for creating the computational grids for Aero-S structural dynamics solvers.

The integration of these components (implemented mostly in the bash scripts and M2C) is reflected in

1.  Simultaneous execution of multiple parallel FSI simulations (\~100 CPU cores each) on a computer cluster, without user intervention.

2.  In each FSI simulation: Real-time, two-way coupling between M2C and Aero-S, with data exchange at every time step.

3.  Format conversions and data handling: SOFICS manages all necessary data transformations and file exchanges between different software components. This includes mesh generation and adaptation using Gmsh, conversion of Gmsh output to the mesh format required by Aero-S, extraction of structural response data from Aero-S output, and computation of constraint and objective function values for Dakota\'s optimization routines.

The integration achieved by SOFICS uniquely addresses structural optimization problems that require fluid-structure coupled simulations. This capability cannot be achieved without any of the four software components. Furthermore, replacing M2C (or another component) with another solver would not be easy and would necessitate substantial and intricate modifications to the underlying solver implementation and the workflow within SOFICS.

Finally, M2C has a significantly broader scope than simulating detonation-induced fluid flows. Many components of the code are
unrelated to this specific application. For example, it includes an incompressible flow solver (e.g., TimeIntegratorSemiImp.h/cpp,
IncompressibleOperator.h/cpp), a laser radiation solver (e.g., LaserAbsorptionSolver.h/cpp), models and solvers for ionization (e.g.,
AtomicIonizationData.h/cpp, NonIdealSahaEquationSolver.h/cpp, IonizationOperator.h/cpp), viscosity (e.g., ViscosityOperator.h/cpp, ViscoFcn.h), and hyperelasticity (e.g., HyperelasticityOperator.h/cpp, ReferenceMapOperator.h/cpp), among others. Of the test cases provided in the source repository (<https://github.com/kevinwgy/m2c/tree/main/Tests>), only about one-third are (directly or indirectly) related to detonations.

Again, we sincerely appreciate the reviewer's comments and questions. In revising the manuscript, we will clarify the roles of the various software components and highlight the integration efforts undertaken to address a specific and challenging optimization problem.

## Comment #2

***Could you avoid hard-copying external libraries into the repo, since they already have their own repositories? An installation script would help automate downloading these dependencies.***

**Response:** We appreciate the reviewer's insightful comment, and we agree that careful dependency management is essential. All three solvers, M2C, Aero-S, and Dakota, are developed and managed by independent research teams of varying sizes, and all are continuously evolving. Dakota is a general-purpose optimization and uncertainty quantification tool that is backed by a relatively large team with a mature release process. As a result, the latest releases of Dakota are often backward compatible. In contrast, M2C and Aero-S are being developed by smaller research teams focused on introducing novel methodologies, which increases the likelihood that new changes may not be fully backward compatible.

That said, we think there are probably three different approaches to managing these three components within SOFICS:

1. Provide links to the latest upstream versions,

2. Provide links to external repositories that host frozen (version-constant) snapshots of Dakota, Aero-S, and M2C,

3. Provide local snapshots within the package itself.

Our initial reasoning for including local snapshots was that approaches (1) and (2) can increase the risk of crashes during the structural optimization runs, whereas (3) lets SOFICS build known, good configurations that, in turn, ensures reproducibility. By default, SOFICS builds the local snapshots of Aero-S and M2C and uses the resulting executables for the coupled FSI analysis. For users who maintain their own modified builds, SOFICS also supports overriding the defaults through the environment variables:

1.  AEROS_EXE=/path/to/aeros

2.  M2C_EXE=/path/to/m2c

When these variables are set, SOFICS skips the local builds and invokes the user-provided executables.

## Comment #3

***Why are you linking gmsh to the specific 4_13_1 version?***

**Response:** We appreciate the reviewer's insightful comment and are grateful for highlighting a potential gap in the description of the dependencies of SOFICS. While SOFICS does not explicitly depend on Gmsh, we agree that a clarification in the README file is necessary.

In response to your concern, we have corrected the links to point to the Gmsh and Dakota home pages instead of their GitLab and GitHub repositories, respectively. We have also included a note clarifying that SOFICS does not require a specific version of Gmsh, while indicating the version used during development and testing for reproducibility. The relevant changes are as follows:

Old:

**[Updated:]{.underline}**

## Comment #4

***In the copied aeros library, there's a CTestTestfile.cmake, is there a reason to keep this? Using the original repositories directly would be best.***

**Response:** We appreciate your observation. This file is a CMake/CTest file generated to drive Aero-S's internal test suite. You are correct in pointing out that it is not required to compile or run Aero-S within SOFICS. In the next revision we will remove this file from the repository, and add an appropriate gitignore rule to prevent any future issues.

## Comment #5

***Could you add an example in the README showing what the software does? Please include some outputs of an FSI coupling.***

**Response:** We sincerely appreciate your kind remarks and insightful comments. We agree that an example case depicting the outputs of a coupled fluid-structure interaction problem in the repository's README file would be prudent. We also concur that the example should be designed to be reproducible, run in a short amount of time, and require relatively modest computational resources, as not all users have access to high-performance computing facilities.

In response to your concern, we plan to add another example case, where we optimize a cantilever beam subjected to a high-pressure load from a propagating shock wave in a partially confined chamber, under "examples/blast_panel_optimization".

In this setup, illustrated in the figure below, the cantilever beam will be modeled using shell finite elements. This approach allows for the use of a relatively coarse finite volume mesh for the fluid dynamics simulation, thereby reducing computational costs. The beam geometry will be parameterized by its section thicknesses, divided into three segments to keep the number of design variables small. The resulting optimization study is expected to yield results similar to those shown in the accompanying figure.

**Figure**: Numerical setup. Note: these parameters are not final and are subject to change.

**Figure**: Example FSI snapshots from optimization of a solid panel subjected to shock impulse.
