### 2025-08-29 #8494 @anarkh97
Hi @kyleniemeyer,

I would like to add the corresponding author of the paper, @kevinwgy, to this discussion so that they can also follow along and contribute as needed. Please let me know if this is the right way to do so, or if there’s a different procedure I should follow.
