### 2025-09-01 #8494 @kevinwgy
Dear Alban (@albansouche) and Kyle (@kyleniemeyer),

I am the corresponding author of this manuscript and a co-developer of SOFICS. Thank you both for reviewing our manuscript and providing the valuable comments and suggestions.

Aditya and I worked together on the response letter, which he submitted three weeks ago. To briefly summarize one key point: the contribution of SOFICS is beyond the bash script. Specifically, M2C was extended to couple with the structural dynamics solver (Aero-S) and the optimizer (Dakota), and it was customized to properly (and efficiently) simulate detonations. While M2C itself has a broader scope --- it is the main research code of our group, with many components not directly used by SOFICS --- its extension and customization are an important part of this manuscript’s contribution. We are happy to clarify this more explicitly when revising the manuscript. (Additional details were provided in the response letter Aditya submitted.)

As this is our first submission, we are not very familiar with review process of JOSS. Our current understanding is that we are waiting for feedback from the editor and reviewers. If that is correct, we greatly appreciate any feedback you can provide. Aditya is a Ph.D. student planning to graduate at the end of this semester (Nov./Dec.). Once we get your comments and suggestions, we will do our best to revise the manuscript promptly and carefully. If, instead, you are waiting to receive a revised submission from us, please kindly let us know.

Thanks again for your time and consideration.

Kevin Wang
Department of Aerospace and Ocean Engineering
Virginia Tech
