### 2025-09-04 #8494 @albansouche
Thank you for the clarifications @anarkh97 and @kevinwgy. I appreciate your detailed answer. However, after spending some time trying to properly install and build SOFICS, I am still not there. In addition, the examples are tailored for cluster runs, but there is no simple, clearly documented example that can be run locally.
As a reviewer, I need to complete the checklist, and, as for now, I can not consider the following items to be adequate:
- **Installation instructions**: the steps are not straightforward and there is no automation in the process.
- **Example usage**: A minimal illustrative example to run on a local machine is missing.
- **Automated tests**: no basic validation tests are available.

I did raise these concerns in my previous comments, but the repository update (only four lines changed) does not address them. For this tool to be usable by researchers outside your group, and to enable me to properly review it in the meantime, I believe addressing these points is essential.
