### 2025-09-07 #8494 @kevinwgy
Thank you @albansouche and @kyleniemeyer for your feedback! We will improve the compilation process and add a few example & code verification tests, as both of you suggested.
