### 2026-02-19 #8494 @anarkh97
Dear @albansouche and @kyleniemeyer,

Thank you very much for your patience, and apologies for the delay. Since our last correspondence, we have made substantial revisions to the SOFICS toolkit. We sincerely appreciate your constructive suggestions for improving the quality, clarity, and usability of this repository. Below we have summarized the changes made in response to these suggestions.
We hope these changes adequately address the concerns raised. We would be happy to discuss any of these points further or make additional modifications as needed.

Regards,
@kevinwgy and @anarkh97

## 1. Project Structure and Dependency Management

As pointed out by @albansouche, local snapshots (i.e., hard-copies) of M2C and Aero-S were initially packaged with SOFICS, which locks the two repositories to certain versions, possibly creating issues in future. We sincerely appreciate @alansouche for noting it in this review.

Subsequently, these local snapshots have been converted to git submodules. Specifically, commits `1484373` from `https://github.com/anarkh97/m2c` and `078fb25` from `https://github.com/anarkh97/aero-s-foam`.

The `.gitmodules` file now references external forks of `https://github.com/anarkh97/m2c` and `https://github.com/anarkh97/aero-s-foam`, ensuring that these solvers are maintained in their own repositories while SOFICS pins specific, tested versions.

Additionally, the project directory has been reorganized for clarity. The former flat `src` directory has been separated into a `scripts` directory (containing the Bash scripts used for streamlining the optimization process) and a `tools` directory (containing the C++ utilities gmsh2aeros and aeros2dakota), each with dedicated CMake build configurations.

## 2. Build Automation and Installation Instructions

A new dependency installation guide (`doc/INSTALL_DEPENDENCIES.md`) has been added with step-by-step instructions for Ubuntu/Debian systems, covering all required and recommended libraries (MPI, Eigen3, BLAS/LAPACK, MUMPS, ScaLAPACK, SPOOLES, etc.), along with a quick setup guide for convenience.

A top-level `build.sh` script has been added that handles CMake configuration and compilation in a single command. After initializing the submodules, the entire toolkit—including M2C, Aero-S, and the SOFICS C++ utilities—can be built with

```sh
./build.sh
```

The final binaries will be located in `/path/to/sofics/build/sofics/bin`.

The build script supports configurable options such as build type, installation prefix, selective component builds, and parallel compilation. The `README.md` includes verification steps with expected output screenshots for each build component.

Additionally, a demonstration video (`doc/demo.mp4`) has also been added. This video covers the workflow for using SOFICS—including cloning, building, and running the 1D piston example. The demonstration was conducted on @anarkh97’s laptop that uses the Windows Subsystem for Linux (WSL). Therefore, it will be helpful for a broad user base, including those who do not have immediate access to high-performance computing (HPC) resources.

## 3. Examples and Documentation
Two new examples have been added to the repository:

- 1D Gas Piston (`examples/1d_gas_piston/`): A new, self-contained example featuring a spring-mass piston system subjected to repeated shock loading inside a one-dimensional shock tube. The example includes a custom piston solver written in C++ (with its own CMake build and input parser), M2C input templates for the compressible fluid domain, configuration files, a Dakota input file for genetic-algorithm-based optimization, and user-defined pre/post-processing scripts. It demonstrates the full SOFICS workflow: Dakota   supplies the design parameters (spring stiffness and mass, derived from two continuous design variables), the pre-processor computes the equivalent spring properties and prepares the input files, M2C and the piston solver run concurrently with two-way coupling through MPI (exchanging interface displacements and fluid loads at every time step), and the post-processor extracts velocity- and displacement-based response functions through the `aeros2dakota` utility for Dakota's optimization loop. Importantly, this example is designed to be lightweight—each coupled simulation uses only 4 MPI processes (3 for M2C and 1 for the piston solver), and the fluid mesh is one-dimensional (single cell in the transverse directions). With an evaluation concurrency of 2, full optimization can be executed on a standard personal workstation with 8 available cores, without requiring access to an HPC cluster.

- Shock Panel (`examples/shock_panel/`): This example was added to depict a demonstrative case showing the outputs of a coupled FSI simulation. Compared to the explosion containment case—which involves the Jones-Wilkins-Lee equation of state, spherical detonation-induced loading, and a three-dimensional ellipsoidal containment structure—the shock panel is a simpler problem where a cantilever beam of variable thickness is subjected to a planar shock wave in an ideal gas. The underlying physics (shock-beam interaction, beam deflection and stress development) are well-studied and the results—tip displacement, von Mises stress, and structural mass as functions of the beam's section thicknesses—are straightforward to interpret and verify. At the same time, this example is more involved than the 1D gas piston case because it exercises the complete SOFICS component chain: `GMsh` generates the structural mesh from a parameterized `.geo` file (with three design variables controlling the tip, mid-chord, and root thicknesses), `gmsh2aeros` converts it to Aero-S format, M2C and Aero-S run concurrently with two-way coupling through the embedded boundary method, `aeros2dakota` extracts the response functions, and Dakota drives the genetic-algorithm optimization loop.

## 4. Scope

As discussed earlier in this thread, the development of SOFICS involved extending M2C with capabilities specifically required for this framework—including two-way fluid-structure coupling with Aero-S, the Jones-Wilkins-Lee equation of state for detonation product thermodynamics, and the UserDataFile input option for initializing detonation scenarios. In addition to these solver extensions, the SOFICS toolkit include the C++ utilities gmsh2aeros (mesh format conversion) and aeros2dakota (result conversion), a modular CMake build system that compiles both solvers and the utilities from a single entry point, and the Bash scripts that manage the end-to-end optimization workflow—from Dakota-driven parameterization and mesh generation through parallel coupled simulations to post-processing of objective and constraint functions.Together, these components form an integrated framework that enables structural optimization under detonation-induced loads.

## 5. Future Plans
We plan to continue developing SOFICS improving it from baseline FSI-based optimization toolkit to SOFICS v2.0 which will support a novel adaptive multi-fidelity optimization framework that is currently under testing.

In the baseline workflow (i.e., SOFICS v1.0) described above, every design evaluation invokes a high-fidelity coupled M2C–Aero-S simulation, which requires significant computational resources (e.g., 64 cores per evaluation in the cantilever panel case). For large-scale optimization problems involving hundreds or thousands of evaluations, this cost can become prohibitive.

To address this, the SOFICS toolkit already supports three evaluation modes—TRUE, APPROX, and ERROR—selectable on a per-evaluation basis through a state variable passed by Dakota. In the APPROX mode, the expensive M2C fluid solver is replaced by a faster approximate fluid solver, which constructs its estimate using the surface meshes, pressure time histories, and design parameters from previously evaluated neighboring designs. SOFICS manages this data exchange automatically. For each approximate evaluation, a metadata file is generated that contains the paths and parameters of the nearest-neighbor evaluations, allowing the approximate solver to interpolate or extrapolate the fluid loading on the structure. In the ERROR mode, SOFICS runs a dedicated error estimation step that computes the discrepancy between the approximate and true solutions, enabling the optimizer to adaptively decide when a full-fidelity evaluation is needed and when the approximation is sufficiently accurate.
This framework is already implemented in the SOFICS v2.0 scripts and is being tested in coordination with an external optimization tool that will be incorporated into SOFICS. We plan to add a documented example demonstrating this multi-fidelity capability once the validation is complete.
